uORB.h reduce orb_metadata field sizes

+++ src/modules/uORB/uORB.h
@@ -49,8 +49,8 @@
  */
 struct orb_metadata {
 	const char *o_name;		/**< unique object name */
-	const size_t o_size;		/**< object size */
-	const size_t o_size_no_padding;	/**< object size w/o padding at the end (for logger) */
+	const uint16_t o_size;		/**< object size */
+	const uint16_t o_size_no_padding;	/**< object size w/o padding at the end (for logger) */
 	const char *o_fields;		/**< semicolon separated list of fields (with type) */
 };
